/roomlist 테스트 및 디버깅 1차

+++ code/carpool-backend/routes/roomlist.js
@@ -1,6 +1,8 @@
 const express = require('express');
+const { connect } = require('./auth');
 const router = express.Router();
 var DB = require('../db/initiate').connection;
+var pool = require('../db/initiate').pool;
 // 미구현된 부분은 TODO : task의 형식으로 달았다.
 
 // GET /roomlist
@@ -53,11 +55,11 @@ router.get('/', function (req, res, next) {
                     console.log(rows)
 
                     res.json(rows) // 3. 사용자에게 json 형식으로 출력하기
-                    res.status(200).end()
-                })
+                    res.status(200);
+                });
             }
-            connection.release();
-        })
+        });
+        connection.release();
     }
 });
 
@@ -87,155 +89,161 @@ router.post('/', (req, res, next) => {
             // 1. MySql query문 만들기
             //`'${car_type}','${depart_place}','${arrive_place}','${depart_time}','${arrive_time}'` 와 같은 형태를 취한다
             //시간: 2020-07-29 14:10:23 형태를 한다
-            var sql = `INSERT INTO carpooldb.roominfos (car_type, depart_place, arrive_place, depart_time, arrive_time,current_headcount, total_headcount, curreunt_carrier_num, total_carrier_num, isConfirm, confirm_time) VALUES(?,?,?,?,?,0,0,0,0,0,NOW())`;
+            var sql = `INSERT INTO carpooldb.roominfos (car_type, depart_place, arrive_place, depart_time, arrive_time,current_headcount, total_headcount, current_carrier_num, total_carrier_num, isConfirm, confirm_time) VALUES(?,?,?,?,?,0,0,0,0,0,NOW())`;
 
             // 2. DB.query로 보내서 출력 가져오기
             DB((err1, connection) => {
                 if (!err1) {
                     connection.query(sql, [car_type, depart_place, arrive_place, depart_time, arrive_time], (err2, rows, fields) => {
                         if (err2) throw err2;
-
-                    })
+                    });
                 }
-            })
+            });
             connection.release();
-            res.status(200)
-            res.redirect('/')
-            res.end()
+            res.status(200);
         }
     }
 });
 
 // PUT /roomlist
-router.put('/', (req, res) => {
+// 입력을 1개씩 못 받는 에러가 있음.
+router.put('/', (req, res, next) => {
     // TODO : 로그인 에러
     if (req.user == undefined) {
         next(new Error('PUT /roomlist error:0'));
     } else {
         // UPDATE tablename SET column1 = value1, column2 = value2, ... WHERE condition
-        res.send('PUT /roomlist');
-
-        console.log(req.body);
         var regExp = /^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}$/;
-        var updateData = [req.body.car_type, req.body.depart_place, req.body.arrive_place, req.body.depart_time, req.body.arrive_place, req.body.current_headcount, req.body.total_headcount, eq.body.current_carrier_num, req.body.total_carrier_num, req.body.isConfirm, req.body.isConfirm_time];
-
+        var updateData = [req.body.car_type, req.body.depart_place, req.body.arrive_place, req.body.depart_time, req.body.arrive_time, req.body.current_headcount, req.body.total_headcount, req.body.current_carrier_num, req.body.total_carrier_num, req.body.isConfirm, req.body.confirm_time];
         // PUT /roomlist 입력 값 에러 : 2
-        if (updateQueryArray.every(undefined) || !regExp.test(updateData[3]) || !regExp.test(updateData[4])) {
+        // 모든 값이 undefined이거나
+        // depart_time이 defined일 때 정규표현식에서 틀리거나
+        // arrive_time이 defined일 때 정규표현식에서 틀리거나
+        // isConfirmTime이 defined일 때 정규표현식에서 틀리거나.
+        if (updateData.every((value)=>{return value==undefined?1:0;}) || (updateData[3] != undefined && !regExp.test(updateData[3])) || (updateData[4] != undefined && !regExp.test(updateData[4])) || (updateData[10] != undefined && !regExp.test(updateData[10])) || req.body.id == undefined) {
             next(new Error('PUT /roomlist error:1'));
         } else {
             // query 생성
             var updateQuery = "UPDATE carpoolDB.roominfos SET";
-            var roomCol = ["car_type", "depart_place", "arrive_place", "depart_time", "depart_place", "current_headcount", "current_carrier_num", "total_carrier_num", "isConfirm", "isConfirm_time"];
+            var roomCol = ["car_type", "depart_place", "arrive_place", "depart_time", "arrive_time", "current_headcount", "total_headcount", "current_carrier_num", "total_carrier_num", "isConfirm", "confirm_time"];
+            console.log(roomCol, '\n', updateData);
             var notNULLcolumn = new Array();
             var j = 0;
+            console.log(updateData.length);
             for (var i = 0; i < updateData.length; i++) {
                 if (updateData[i] != undefined) {
                     updateQuery = updateQuery + " " + roomCol[i] + "=?,";
-                    notNULLcolumn[j] = updateQueryArray[i]; j++;
+                    notNULLcolumn[j] = updateData[i]; j++;
                 }
             }
-            updateQuery = updateQuery.slice(0, -1) + `WHERE id=?`
+            updateQuery = updateQuery.slice(0, -1) + ` WHERE id=?`
             notNULLcolumn[j] = req.body.id;
+            console.log(updateQuery, notNULLcolumn);
 
             // 값 업데이트
             DB((err, connection) => {
                 if (err) {
                     // TODO : DB에 접근 못 할때
                     console.log("PUT /roomlist error : 서버 이용자가 너무 많습니다.");
+                    connection.release();
                     next(new Error('PUT /roomlist error:3'));
                 } else {
                     connection.query(updateQuery, notNULLcolumn, (sqlErr) => {
                         if (sqlErr) {
                             // TODO : sql 내부 에러 처리
                             console.log("PUT /roomlist error : SQL 내부 에러. query를 확인해 주세요.");
+                            next(new Error('PUT /roomlist error:4'));
                         } else {
                             console.log("업데이트 완료");
-                            res.status(200)
-                            res.redirect('/login');
+                            res.status(200);
                         }
-                    })
+                    });
+                    connection.release();
                 }
-                connection.release();
             });
         }
-
-        res.end()
     }
 });
 
 // DELETE /roomlist
-router.delete('/', (req, res) => {
+// pass
+router.delete('/:roomID', (req, res, next) => {
     // TODO : 로그인 에러
     if (req.user == undefined) {
         next(new Error('DELETE /roomlist error:0'));
     } else {
         // DELETE FROM tablename WHERE condition;
         // 값 삭제
-        var deleteQuery = `DELETE FROM carpoolDB.roominfos WHERE id=?`;
+        var deleteQuery = `DELETE FROM carpoolDB.roominfos WHERE id=?; DELETE FROM carpoolDB.users_and_rooms_infos WHERE roomId = ?`;
         DB((err, connection) => {
             if (err) {
                 // TODO : DB에 접근 못 할때
                 console.log("DELETE /roomlist error : 서버 이용자가 너무 많습니다.");
+                connection.release();
                 next(new Error('DELETE /roomlist error:3'));
             } else {
-                connection.query(deleteQuery, [req.body.id], (sqlErr) => {
+                connection.query(deleteQuery, [req.params.roomID, req.params.roomID], (sqlErr) => {
                     if (sqlErr) {
                         // TODO : sql 내부 에러 처리
                         console.log("DELETE /roomlist error : SQL 내부 에러. query를 확인해 주세요. 해당하는 방이 없습니다.");
+                        next(new Error('DELETE /roomlist error:4'));
                     } else {
                         console.log("삭제 완료");
-                        res.status(200)
-                        res.redirect('/login');
+                        res.status(200);
                     }
-                })
+                });
+                connection.release();
             }
-            connection.release();
         });
-
-        res.end();
     }
 });
 
-
-// GET /roomlist/userid?id=
-router.get('/userid', function (req, res, next) {
+// GET /roomlist/userid/:userID
+// pass
+router.get('/userid/:userID', function (req, res, next) {
     // TODO : 로그인 에러
+    console.log(req.params.userID);
     if (req.user == undefined) {
         next(new Error('GET /roomlist error:0'));
     } else {
         // admin이 아닌데 다른 id로 접근하는 경우에는 에러
-        if (req.user.isAdmin == 0 && req.query.id != req.user.id) {
+        if (req.user.isAdmin == 0 && req.params.userID != req.user.id) {
             // TODO : 접근 권한 오류
             next(new Error('GET /roomlist error:2'));
         } else {
             // userid가 속한 방에 대한 정보 출력, ./db/testquery 파일 참고.
-            var belongQuery = `SELECT roominfos.id, car_type, depart_place, arrive_place, depart_time, arrive_time, current_headcount, total_headcount, curreunt_carrier_num, total_carrier_num, isConfirm, confirm_time
-        FROM carpooldb.roominfos INNER JOIN carpooldb.users_and_rooms_infos ON roominfos.id = users_and_rooms_infos.roomID WHERE users_and_rooms_infos.userid = ? ORDER BY depart_time ASC;`;
+            var belongQuery = `SELECT roominfos.id, car_type, depart_place, arrive_place, depart_time, arrive_time, current_headcount, total_headcount,
+            current_carrier_num, total_carrier_num, isConfirm, confirm_time FROM carpooldb.roominfos INNER JOIN carpooldb.users_and_rooms_infos
+            ON roominfos.id = users_and_rooms_infos.roomID WHERE users_and_rooms_infos.userid = ? ORDER BY depart_time ASC;`;
 
             DB((err, connection) => {
                 if (err) {
                     // TODO : DB에 접근 못 할 때
                     console.log("GET /roomlist/userid?id= error : 서버 이용자가 너무 많습니다.");
+                    connection.release();
                     next(new Error('GET /roomlist error:3'));
                 } else {
-                    connection.query(belongQuery, [req.query.id], (err, result) => {
+                    connection.query(belongQuery, [req.params.userID], (err, result) => {
+                        console.log(belongQuery);
                         if (err) {
                             // TODO : sql 내부 에러 처리
                             console.log("GET /roomlist/userid?id= error : SQL 내부 에러. query를 확인해 주세요.");
+                            next(new Error('GET /roomlist error:4'));
                         } else{
                             console.log("id에 해당하는 사람이 속해있는 방의 정보 출력", result);
                             res.json(result);
-                            res.status(200).end();
+                            res.status(200);
                         }
                     });
+                    connection.release();
                 }
-                connection.release();
             });
         }
     }
-});
+}); 
 
 // POST /roomlist/userid
+// pass
 router.post('/userid', function (req, res, next) {
     // TODO : 로그인 에러
     if (req.user == undefined) {
@@ -249,59 +257,64 @@ router.post('/userid', function (req, res, next) {
                                             AND EXISTS (SELECT roominfos.id FROM carpooldb.roominfos WHERE roominfos.id = ?)
                                             AND NOT EXISTS (SELECT * FROM carpooldb.users_and_rooms_infos WHERE userid = ? AND roomid = ?)
                                         LIMIT 1);`;
-        /* var userID = req.body.userID;
-        var roomID = req.body.roomID; */
-        var userID = 1;
-        var roomID = 5;
+        var userID = req.body.userID;
+        var roomID = req.body.roomID;
         DB((err, connection) => {
             if (err) {
                 // TODO : DB에 접근 못 할 때
                 console.log("POST /roomlist/userid error : 서버 이용자가 너무 많습니다.");
-                next(new Error('POST /roomlist/userid error:0'));
+                connection.release();
+                next(new Error('POST /roomlist/userid error:3'));
             } else {
                 connection.query(addUsersRoomQuery, [userID, roomID, userID, roomID, userID, roomID], (err, result) => {
                     if (err) {
                         // TODO : sql 내부 에러 처리
                         console.log("POST /roomlist/userid error : SQL 내부 에러. query를 확인해 주세요.");
+                        next(new Error('POST /roomlist/userid error:4'));
                     } else{
-                        res.json(result);
-                        res.status(200).end();
+                        res.status(200);
                     }
                 });
+                connection.release();
             }
-            connection.release();
         });
     }
 });
-// DELETE /roomlist/userid
-router.delete('/userid', (req, res) => {
+
+// DELETE /roomlist/userid/:userID/:roomID
+// pass
+router.delete('/userid/:userID/:roomID', (req, res, next) => {
     // TODO : 로그인 에러
     if (req.user == undefined) {
         next(new Error('DELETE /roomlist/userid error:0'));
     } else {
-        // query
-        var deleteQuery = `DELETE FROM carpooldb.users_and_rooms_infos WHERE userid = ? AND roomid = ?;`;
-        DB((err, connection) => {
-            if (err) {
-                // TODO : DB에 접근 못 할때
-                console.log("DELETE /roomlist/userid error : 서버 이용자가 너무 많습니다.");
-                next(new Error('DELETE /roomlist/userid error:3'));
-            } else {
-                connection.query(deleteQuery, [req.body.userID, req.body.roomID], (sqlErr) => {
-                    if (sqlErr) {
-                        // TODO : sql 내부 에러 처리
-                        console.log("DELETE /roomlist/userid error : SQL 내부 에러. query를 확인해 주세요.");
-                    } else {
-                        console.log("삭제 완료");
-                        res.status(200)
-                        res.redirect('/login');
-                    }
-                });
-            }
-            connection.release();
-        });
+        if (req.user.isAdmin == 0 && req.params.userID != req.user.id) {    
+            // TODO : 접근 권한 오류
+            next(new Error('DELETE /roomlist/userid/:userID/:roomID error:2'));
+        } else {
+            var deleteQuery = `DELETE FROM carpooldb.users_and_rooms_infos WHERE userid = ? AND roomid = ?;`;
+            DB((err, connection) => {
+                if (err) {
+                    // TODO : DB에 접근 못 할때
+                    console.log("DELETE /roomlist/userid error : 서버 이용자가 너무 많습니다.");
+                    connection.release();
+                    next(new Error('DELETE /roomlist/userid error:3'));
+                } else {
+                    connection.query(deleteQuery, [req.params.userID, req.params.roomID], (sqlErr) => {
+                        if (sqlErr) {
+                            // TODO : sql 내부 에러 처리
+                            console.log("DELETE /roomlist/userid error : SQL 내부 에러. query를 확인해 주세요.");
+                            next(new Error('DELETE /roomlist/userid error:4'));
+                        } else {
+                            console.log("삭제 완료");
+                            res.status(200);
+                        }
+                    });
+                    connection.release();
+                }
+            });
+        }
     }
-    res.send('DELETE /roomlist/userid');
 });
 
 // 오류 처리기
@@ -309,6 +322,7 @@ router.delete('/userid', (req, res) => {
 // 1 : TODO : 입력값 에러
 // 2 : TODO : 접근 에러
 // 3 : TODO : 동시에 너무 많은 접속이 있을 때
+// 4 : TODO : query 에러
 router.use((err, req, res, next) => {
     res.json({ message: err.message });
 })
